+++ packages/hash/frontend/src/pages/[account-slug]/types/entity-type/[entity-type-id].page.tsx
@@ -4,27 +4,26 @@ import {
   extractVersion,
   PropertyTypeReference,
   ValueOrArray,
-  VersionedUri,
 } from "@blockprotocol/type-system-web";
-import { faAsterisk, faPlus } from "@fortawesome/free-solid-svg-icons";
+import { faAsterisk } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@hashintel/hash-design-system/fontawesome-icon";
 import { Box, Container, Typography } from "@mui/material";
 import { Buffer } from "buffer/";
 import { useRouter } from "next/router";
 import { useMemo } from "react";
-import { FormProvider, useForm, useWatch } from "react-hook-form";
+import { FormProvider, useForm } from "react-hook-form";
 import { FRONTEND_URL } from "../../../../lib/config";
 import { getPlainLayout, NextPageWithLayout } from "../../../../shared/layout";
 import { TopContextBar } from "../../../shared/top-context-bar";
 import { HashOntologyIcon } from "../../shared/hash-ontology-icon";
 import { OntologyChip } from "../../shared/ontology-chip";
 import { EditBar } from "./edit-bar";
+import { EntityTypeTabs } from "./entity-type-tabs";
 import {
   EntityTypeEditorForm,
   EntityTypeEditorPropertyData,
 } from "./form-types";
 import { PropertyListCard } from "./property-list-card";
-import { TabButton } from "./tab-button";
 import { useEntityType } from "./use-entity-type";
 import {
   PropertyTypesContext,
@@ -99,8 +98,7 @@ const Page: NextPageWithLayout = () => {
     defaultValues: { properties: [] },
   });
 
-  const { handleSubmit: wrapHandleSubmit, reset, control } = formMethods;
-  const propertiesCount = useWatch({ control, name: "properties.length" });
+  const { handleSubmit: wrapHandleSubmit, reset } = formMethods;
 
   const [remoteEntityType, updateEntityType, publishDraft] = useEntityType(
     baseEntityTypeUri,
@@ -260,56 +258,7 @@ const Page: NextPageWithLayout = () => {
                   />
                   {entityType.title}
                 </Typography>
-                <Box display="flex">
-                  <TabButton
-                    href={extractBaseUri(entityType.$id as VersionedUri)}
-                    sx={(theme) => ({
-                      borderBottomColor: theme.palette.blue[60],
-                      color: theme.palette.blue[70],
-                    })}
-                  >
-                    <Typography
-                      variant="smallTextLabels"
-                      sx={{ fontWeight: 500 }}
-                    >
-                      Definition
-                    </Typography>
-                    <Box
-                      sx={(theme) => ({
-                        py: 0.25,
-                        px: 1,
-                        backgroundColor: theme.palette.blue[20],
-                        borderRadius: "30px",
-                        ml: 1,
-                      })}
-                    >
-                      <Typography variant="microText" sx={{ fontWeight: 500 }}>
-                        {propertiesCount}
-                      </Typography>
-                    </Box>
-                  </TabButton>
-                  <Box display="flex" ml="auto">
-                    <TabButton
-                      href="#"
-                      sx={(theme) => ({ color: theme.palette.gray[90] })}
-                    >
-                      <Typography
-                        variant="smallTextLabels"
-                        sx={{ fontWeight: 500 }}
-                      >
-                        Create new entity
-                      </Typography>
-                      <FontAwesomeIcon
-                        icon={faPlus}
-                        sx={(theme) => ({
-                          ...theme.typography.smallTextLabels,
-                          color: theme.palette.blue[70],
-                          ml: 1,
-                        })}
-                      />
-                    </TabButton>
-                  </Box>
-                </Box>
+                <EntityTypeTabs entityType={entityType} />
               </Container>
             </Box>
           </Box>
